requirements.txt
fastapi
openai
pydantic
python-dotenv
mysql-connector-python
